LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/shape-margin-expected.txt


PASS Can set 'shape-margin' to CSS-wide keywords
FAIL Can set 'shape-margin' to var() references assert_equals: expected 2 but got 1
FAIL Can set 'shape-margin' to a length Invalid values
FAIL Can set 'shape-margin' to a percent Invalid values
PASS Setting 'shape-margin' to a time throws TypeError
PASS Setting 'shape-margin' to an angle throws TypeError
PASS Setting 'shape-margin' to a flexible length throws TypeError
FAIL Setting 'shape-margin' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'shape-margin' to a URL throws TypeError
PASS Setting 'shape-margin' to a transform throws TypeError
